fcp-microservices/services/sql-generator/tests/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
azure-core==1.30.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
DateTime==5.5
distro==1.9.0
et-xmlfile==1.1.0
filelock==3.15.4
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
msal==1.29.0
msal-extensions==1.2.0
numpy==2.0.0
openai==1.35.12
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.11.0
snowflake-sqlalchemy==1.6.1
sortedcontainers==2.4.0
sql_metadata==2.12.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
tomlkit==0.13.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
zope.interface==6.4.post2